feat: working on config validator
